refactor: move path_hash to shared crate
